Clean up stategen
