+++ beacon-chain/state/stategen/setter_test.go
@@ -4,37 +4,16 @@ import (
 	"context"
 	"testing"
 
+	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/beacon-chain/cache"
 	testDB "github.com/prysmaticlabs/prysm/beacon-chain/db/testing"
+	"github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
 	logTest "github.com/sirupsen/logrus/hooks/test"
 )
-
-func TestSaveState_ColdStateCanBeSaved(t *testing.T) {
-	hook := logTest.NewGlobal()
-	ctx := context.Background()
-	db, _ := testDB.SetupDB(t)
-
-	service := New(db, cache.NewStateSummaryCache())
-	service.slotsPerArchivedPoint = 1
-	beaconState, _ := testutil.DeterministicGenesisState(t, 32)
-
-	// This goes to cold section.
-	slot := uint64(1)
-	require.NoError(t, beaconState.SetSlot(slot))
-	service.finalizedInfo.slot = slot + 1
-
-	r := [32]byte{'a'}
-	require.NoError(t, service.SaveState(ctx, r, beaconState))
-
-	assert.Equal(t, true, service.beaconDB.HasArchivedPoint(ctx, 1), "Did not save cold state")
-	assert.Equal(t, r, service.beaconDB.ArchivedPointRoot(ctx, 1), "Did not get wanted root")
-
-	require.LogsContain(t, hook, "Saved full state on archived point")
-}
 
 func TestSaveState_HotStateCanBeSaved(t *testing.T) {
 	ctx := context.Background()
@@ -94,4 +73,64 @@ func TestState_ForceCheckpoint_SavesStateToDatabase(t *testing.T) {
 
 	// Should not panic with genesis finalized root.
 	assert.NoError(t, svc.ForceCheckpoint(ctx, params.BeaconConfig().ZeroHash[:]))
+}
+
+func TestSaveState_AlreadyHas(t *testing.T) {
+	hook := logTest.NewGlobal()
+	ctx := context.Background()
+	db, _ := testDB.SetupDB(t)
+	service := New(db, cache.NewStateSummaryCache())
+
+	beaconState, _ := testutil.DeterministicGenesisState(t, 32)
+	require.NoError(t, beaconState.SetSlot(params.BeaconConfig().SlotsPerEpoch))
+	r := [32]byte{'A'}
+
+	// Pre cache the hot state.
+	service.hotStateCache.Put(r, beaconState)
+	require.NoError(t, service.saveStateByRoot(ctx, r, beaconState))
+
+	// Should not save the state and state summary.
+	assert.Equal(t, false, service.beaconDB.HasState(ctx, r), "Should not have saved the state")
+	assert.Equal(t, false, service.beaconDB.HasStateSummary(ctx, r), "Should have saved the state summary")
+	require.LogsDoNotContain(t, hook, "Saved full state on epoch boundary")
+}
+
+func TestSaveState_CanSaveOnEpochBoundary(t *testing.T) {
+	ctx := context.Background()
+	db, _ := testDB.SetupDB(t)
+	service := New(db, cache.NewStateSummaryCache())
+
+	beaconState, _ := testutil.DeterministicGenesisState(t, 32)
+	require.NoError(t, beaconState.SetSlot(params.BeaconConfig().SlotsPerEpoch))
+	r := [32]byte{'A'}
+
+	require.NoError(t, service.saveStateByRoot(ctx, r, beaconState))
+
+	// Should save both state and state summary.
+	_, ok, err := service.epochBoundaryStateCache.getByRoot(r)
+	require.NoError(t, err)
+	require.Equal(t, true, ok, "Did not save epoch boundary state")
+	assert.Equal(t, true, service.stateSummaryCache.Has(r), "Should have saved the state summary")
+}
+
+func TestSaveState_NoSaveNotEpochBoundary(t *testing.T) {
+	hook := logTest.NewGlobal()
+	ctx := context.Background()
+	db, _ := testDB.SetupDB(t)
+	service := New(db, cache.NewStateSummaryCache())
+
+	beaconState, _ := testutil.DeterministicGenesisState(t, 32)
+	require.NoError(t, beaconState.SetSlot(params.BeaconConfig().SlotsPerEpoch-1))
+	r := [32]byte{'A'}
+	b := &ethpb.SignedBeaconBlock{Block: &ethpb.BeaconBlock{}}
+	require.NoError(t, db.SaveBlock(ctx, b))
+	gRoot, err := stateutil.BlockRoot(b.Block)
+	require.NoError(t, err)
+	require.NoError(t, db.SaveGenesisBlockRoot(ctx, gRoot))
+	require.NoError(t, service.saveStateByRoot(ctx, r, beaconState))
+
+	// Should only save state summary.
+	assert.Equal(t, false, service.beaconDB.HasState(ctx, r), "Should not have saved the state")
+	assert.Equal(t, true, service.stateSummaryCache.Has(r), "Should have saved the state summary")
+	require.LogsDoNotContain(t, hook, "Saved full state on epoch boundary")
 }
